pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "biasops-policy-marketplace"
version = "0.1.0"
description = "A policy-as-code marketplace for defining, validating, and enforcing bias-detection policies."
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
dependencies = [
    "pydantic>=2.0",
    "jsonschema>=4.20",
    "PyYAML>=6.0",
    "fastapi>=0.110",
    "uvicorn>=0.27",
    "typer>=0.9",
    "rich>=13.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=4.0",
    "pytest-asyncio>=0.23",
    "httpx>=0.27",
    "ruff>=0.3",
    "mypy>=1.8",
]

[project.scripts]
biasops = "biasops.cli:app"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.mypy]
python_version = "3.10"
strict = true
